combining common constraints
