+++ watertap/unit_models/_reverse_osmosis_base.py
@@ -13,7 +13,7 @@
 
 from copy import deepcopy
 from enum import Enum, auto
-from pyomo.environ import Block, NonNegativeReals, Param, Suffix, Var, units as pyunits
+from pyomo.environ import Block, exp, NonNegativeReals, Param, Suffix, Var, units as pyunits
 from pyomo.common.collections import ComponentSet
 from pyomo.common.config import ConfigBlock, ConfigValue, In
 from idaes.core import UnitModelBlockData, useDefault, MaterialBalanceType,\
@@ -260,6 +260,11 @@ class _ReverseOsmosisBaseData(UnitModelBlockData):
         units_meta = \
             self.config.property_package.get_metadata().get_derived_units
 
+        self.nfe = Param(
+            initialize=(len(self.difference_elements)),
+            units=pyunits.dimensionless,
+            doc="Number of finite elements")
+
         """ Unit model variables"""
         self.A_comp = Var(
             self.flowsheet().config.time,
@@ -420,6 +425,223 @@ class _ReverseOsmosisBaseData(UnitModelBlockData):
                 units=pyunits.dimensionless,
                 doc='Concentration polarization modulus')
 
+        # ==========================================================================
+        # Volumetric Recovery rate
+        @self.Constraint(self.flowsheet().config.time)
+        def eq_recovery_vol_phase(b, t):
+            return (b.recovery_vol_phase[t, 'Liq'] ==
+                    b.mixed_permeate[t].flow_vol_phase['Liq'] /
+                    b.feed_side.properties[t, self.first_element].flow_vol_phase['Liq'])
+
+        # ==========================================================================
+        # Mass-based Component Recovery rate
+        @self.Constraint(self.flowsheet().config.time,
+                         self.config.property_package.component_list)
+        def eq_recovery_mass_phase_comp(b, t, j):
+            return (b.recovery_mass_phase_comp[t, 'Liq', j]
+                    * b.feed_side.properties[t, b.first_element].flow_mass_phase_comp['Liq', j] ==
+                    b.mixed_permeate[t].flow_mass_phase_comp['Liq', j])
+
+        @self.Constraint(self.flowsheet().config.time,
+                         self.difference_elements,
+                         self.config.property_package.phase_list,
+                         self.config.property_package.component_list,
+                         doc="Solvent and solute mass flux")
+        def eq_flux_mass(b, t, x, p, j):
+            prop_feed = b.feed_side.properties[t, x]
+            prop_perm = b.permeate_side[t, x]
+            interface = b.feed_side.properties_interface[t, x]
+            comp = self.config.property_package.get_component(j)
+            if comp.is_solvent():
+                return (b.flux_mass_phase_comp[t, x, p, j] == b.A_comp[t, j] * b.dens_solvent
+                        * ((prop_feed.pressure - prop_perm.pressure)
+                           - (interface.pressure_osm - prop_perm.pressure_osm)))
+            elif comp.is_solute():
+                return (b.flux_mass_phase_comp[t, x, p, j] == b.B_comp[t, j]
+                        * (interface.conc_mass_phase_comp[p, j] - prop_perm.conc_mass_phase_comp[p, j]))
+
+        # # ==========================================================================
+        # Concentration polarization
+        @self.feed_side.Constraint(self.flowsheet().config.time,
+                                   self.difference_elements,
+                                   solute_set,
+                                   doc="Concentration polarization")
+        def eq_concentration_polarization(b, t, x, j):
+            bulk = b.properties[t, x]
+            interface = b.properties_interface[t, x]
+            if self.config.concentration_polarization_type == ConcentrationPolarizationType.none:
+                return interface.conc_mass_phase_comp['Liq', j] == \
+                       bulk.conc_mass_phase_comp['Liq', j]
+            elif self.config.concentration_polarization_type == ConcentrationPolarizationType.fixed:
+                return (interface.conc_mass_phase_comp['Liq', j] ==
+                        bulk.conc_mass_phase_comp['Liq', j]
+                        * self.cp_modulus[t, x, j])
+            elif self.config.concentration_polarization_type == ConcentrationPolarizationType.calculated:
+                jw = self.flux_mass_phase_comp[t, x, 'Liq', 'H2O'] / self.dens_solvent
+                js = self.flux_mass_phase_comp[t, x, 'Liq', j]
+                return (interface.conc_mass_phase_comp['Liq', j] ==
+                        bulk.conc_mass_phase_comp['Liq', j] * exp(jw / self.Kf[t, x, j])
+                        - js / jw * (exp(jw / self.Kf[t, x, j]) - 1))
+
+        # # ==========================================================================
+        # Feed and permeate-side isothermal conditions
+        @self.Constraint(self.flowsheet().config.time,
+                         self.length_domain,
+                         doc="Isothermal assumption for permeate")
+        def eq_permeate_isothermal(b, t, x):
+            return b.feed_side.properties[t, x].temperature == \
+                   b.permeate_side[t, x].temperature
+
+        # ==========================================================================
+        # isothermal conditions at permeate outlet
+        @self.Constraint(self.flowsheet().config.time,
+                         doc="Isothermal assumption for permeate out")
+        def eq_permeate_outlet_isothermal(b, t):
+            return b.feed_side.properties[t, b.length_domain.last()].temperature == \
+                   b.mixed_permeate[t].temperature
+
+        # ==========================================================================
+        # isobaric conditions across permeate channel and at permeate outlet
+        @self.Constraint(self.flowsheet().config.time,
+                         self.length_domain,
+                         doc="Isobaric assumption for permeate out")
+        def eq_permeate_outlet_isobaric(b, t, x):
+            return b.permeate_side[t, x].pressure == \
+                   b.mixed_permeate[t].pressure
+
+        # ==========================================================================
+        # Membrane area equation
+        if hasattr(self, 'length') or hasattr(self, 'width'):
+            @self.Constraint(doc="Membrane area")
+            def eq_area(b):
+                return b.area == b.length * b.width
+
+        # # ==========================================================================
+        # Constraints active when MassTransferCoefficient.calculated
+        # Mass transfer coefficient calculation
+        if self.config.mass_transfer_coefficient == MassTransferCoefficient.calculated:
+            @self.Constraint(self.flowsheet().config.time,
+                             self.difference_elements,
+                             solute_set,
+                             doc="Mass transfer coefficient in feed channel")
+            def eq_Kf(b, t, x, j):
+                bulk = b.feed_side.properties[t, x]
+                return (b.Kf[t, x, j] * b.dh ==
+                        bulk.diffus_phase['Liq']  # TODO: add diff coefficient to SW prop and consider multi-components
+                        * b.N_Sh[t, x])
+
+            @self.Constraint(self.flowsheet().config.time,
+                             self.length_domain,
+                             doc="Sherwood number")
+            def eq_N_Sh(b, t, x):
+                return (b.N_Sh[t, x] ==
+                        0.46 * (b.N_Re[t, x] * b.N_Sc[t, x])**0.36)
+
+            @self.Constraint(self.flowsheet().config.time,
+                             self.length_domain,
+                             doc="Schmidt number")
+            def eq_N_Sc(b, t, x):
+                bulk = b.feed_side.properties[t, x]
+                return (b.N_Sc[t, x] * bulk.dens_mass_phase['Liq'] * bulk.diffus_phase['Liq'] ==
+                        bulk.visc_d_phase['Liq'])
+
+        if (self.config.mass_transfer_coefficient == MassTransferCoefficient.calculated
+            or self.config.pressure_change_type == PressureChangeType.calculated):
+            @self.Constraint(doc="Cross-sectional area")
+            def eq_area_cross(b):
+                return b.area_cross == b.channel_height * b.width * b.spacer_porosity
+
+            @self.Constraint(self.flowsheet().config.time,
+                             self.length_domain,
+                             doc="Reynolds number")
+            def eq_N_Re(b, t, x):
+                bulk = b.feed_side.properties[t, x]
+                return (b.N_Re[t, x] * b.area_cross * bulk.visc_d_phase['Liq'] ==
+                        sum(bulk.flow_mass_phase_comp['Liq', j] for j in b.config.property_package.component_list)
+                        * b.dh)
+
+        if self.config.pressure_change_type == PressureChangeType.calculated:
+            ## ==========================================================================
+            # Crossflow velocity
+            @self.Constraint(self.flowsheet().config.time,
+                             self.length_domain,
+                             doc="Crossflow velocity constraint")
+            def eq_velocity(b, t, x):
+                return b.velocity[t, x] * b.area_cross == b.feed_side.properties[t, x].flow_vol_phase['Liq']
+
+            ## ==========================================================================
+            # Darcy friction factor based on eq. S27 in SI for Cost Optimization of Osmotically Assisted Reverse Osmosis
+            # TODO: this relationship for friction factor is specific to a particular spacer geometry. Add alternatives.
+            @self.Constraint(self.flowsheet().config.time,
+                             self.length_domain,
+                             doc="Darcy friction factor constraint")
+            def eq_friction_factor_darcy(b, t, x):
+                return (b.friction_factor_darcy[t, x] - 0.42) * b.N_Re[t, x] == 189.3
+
+            ## ==========================================================================
+            # Pressure change per unit length due to friction
+            # -1/2*f/dh*density*velocity^2
+            @self.Constraint(self.flowsheet().config.time,
+                             self.length_domain,
+                             doc="pressure change per unit length due to friction")
+            def eq_dP_dx(b, t, x):
+                bulk = b.feed_side.properties[t, x]
+                return (b.dP_dx[t, x] * b.dh ==
+                        -0.5 * b.friction_factor_darcy[t, x]
+                        * b.feed_side.properties[t, x].dens_mass_phase['Liq'] * b.velocity[t, x]**2)
+
+        # ==========================================================================
+        # Bulk and interface connections on the feed-side
+        # TEMPERATURE
+        @self.feed_side.Constraint(self.flowsheet().config.time,
+                                   self.difference_elements,
+                                   doc="Temperature at interface")
+        def eq_equal_temp_interface(b, t, x):
+            return b.properties_interface[t, x].temperature == \
+                   b.properties[t, x].temperature
+        # PRESSURE
+        @self.feed_side.Constraint(self.flowsheet().config.time,
+                                   self.difference_elements,
+                                   doc="Pressure at interface")
+        def eq_equal_pressure_interface(b, t, x):
+            return b.properties_interface[t, x].pressure == \
+                   b.properties[t, x].pressure
+        # VOLUMETRIC FLOWRATE
+        @self.feed_side.Constraint(self.flowsheet().config.time,
+                                   self.difference_elements,
+                                   doc="Volumetric flow at interface of inlet")
+        def eq_equal_flow_vol_interface(b, t, x):
+            return b.properties_interface[t, x].flow_vol_phase['Liq'] ==\
+                   b.properties[t, x].flow_vol_phase['Liq']
+
+        @self.Expression(self.flowsheet().config.time,
+                         self.config.property_package.phase_list,
+                         self.config.property_package.component_list,
+                         doc="Average flux expression")
+        def flux_mass_phase_comp_avg(b, t, p, j):
+            return sum(b.flux_mass_phase_comp[t, x, p, j]
+                       for x in self.difference_elements) / self.nfe
+
+    def _add_expressions(self):
+        """
+        Generate expressions for additional results desired for full report
+        """
+
+        solute_set = self.config.property_package.solute_set
+        if hasattr(self, 'N_Re'):
+            @self.Expression(self.flowsheet().config.time,
+                             doc="Average Reynolds Number expression")
+            def N_Re_avg(b, t):
+                return sum(b.N_Re[t, x]
+                           for x in self.length_domain) / self.nfe
+        if hasattr(self, 'Kf'):
+            @self.Expression(self.flowsheet().config.time,
+                             solute_set,
+                             doc="Average mass transfer coefficient expression")
+            def Kf_avg(b, t, j):
+                return sum(b.Kf[t, x, j]
+                           for x in self.difference_elements) / self.nfe
+
     def _get_stream_table_contents(self, time_point=0):
         return create_stream_table_dataframe(
             {
@@ -505,6 +727,7 @@ class _ReverseOsmosisBaseData(UnitModelBlockData):
             state_args_interface_in['flow_mass_phase_comp'][('Liq', j)] *= initialize_guess['cp_modulus']
             state_args_interface_out['flow_mass_phase_comp'][('Liq', j)] *= initialize_guess['cp_modulus']
 
+        # TODO: I think this is what we'd like to do, but IDAES needs to be changed
         state_args_interface = {}
         for t in self.flowsheet().config.time:
             for x in self.length_domain:
@@ -519,6 +742,18 @@ class _ReverseOsmosisBaseData(UnitModelBlockData):
                     else:
                         state_args_tx[k] = (1.-x)*state_args_interface_in[k] + x*state_args_interface_out[k]
                 state_args_interface[t,x] = state_args_tx
+
+        x = 0.5
+        state_args_tx = {}
+        for k in state_args_interface_in:
+            if isinstance(state_args_interface_in[k], dict):
+                if k not in state_args_tx:
+                    state_args_tx[k] = {}
+                for index in state_args_interface_in[k]:
+                    state_args_tx[k][index] = (1.-x)*state_args_interface_in[k][index] + x*state_args_interface_out[k][index]
+            else:
+                state_args_tx[k] = (1.-x)*state_args_interface_in[k] + x*state_args_interface_out[k]
+        state_args_interface = state_args_tx
 
         return {'feed_side' : state_args,
                 'retentate' : state_args_retentate,
